pyproject.toml
[project]
name = "paulblish"
version = "0.1.0"
description = "Turn your markdown directory into a static site"
authors = [{ name = "Paul Hallett", email = "paulandrewhallett@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-frontmatter>=1.1",
    "markdown-it-py>=3.0",
    "jinja2>=3.1",
    "pygments>=2.18",
    "click>=8.1",
    "mdit-py-plugins>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/phalt/paulblish"
Issues = "https://github.com/phalt/paulblish/issues"

[project.scripts]
pb = "paulblish.cli:main"

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.9"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["paulblish"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
